fix: some errors and add response failure to readme

+++ domain/login.go
@@ -7,7 +7,7 @@ var (
 )
 
 type Login struct {
-	Cpf    int64
+	Cpf    int
 	Secret string
 }
