+++ http/createAccount.go
@@ -11,7 +11,7 @@ import (
 
 type AccountRequest struct {
 	Name    string  `json:"name"`
-	Cpf     int64   `json:"cpf"`
+	Cpf     int     `json:"cpf"`
 	Secret  string  `json:"secret"`
 	Balance float64 `json:"balance"`
 }
@@ -19,7 +19,7 @@ type AccountRequest struct {
 type AccountResponse struct {
 	Id        string    `json:"id"`
 	Name      string    `json:"name"`
-	Cpf       int64     `json:"cpf"`
+	Cpf       int       `json:"cpf"`
 	Balance   float64   `json:"balance"`
 	CreatedAt time.Time `json:"createdAt"`
 }
